requirements/devlopment.txt
-r common.txt
South==0.7.3
django-debug-toolbar==0.9.4
